Move ViewDefinition to presto-analyzer
Moving ViewDefinition to presto-analyzer.
ViewDefinition relies on a few classes from Guava, which is an issue
as this move was leading to presto-spi depending on guava.
Hence guava related classes have been removed from ViewDefinition.
